system_services/hailo-depth/requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
pillow>=9.0
numpy>=1.21
opencv-python>=4.5
requests>=2.28
matplotlib>=3.5
